requirements.txt
boto3==1.16.59
botocore==1.19.59
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.6
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gdown==3.12.2
idna==2.10
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
m3u8==0.8.0
MarkupSafe==1.1.1
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
pymongo==3.11.2
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.3.4
shortuuid==1.0.1
six==1.15.0
threaded==4.1.0
tqdm==4.56.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
